% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBV_functions.R
\name{cor_var}
\alias{cor_var}
\title{simulating correlated variable}
\usage{
cor_var(x, cor)
}
\arguments{
\item{x}{Vector of values to which the other vector needs to be correlated}

\item{cor}{correlation}
}
\value{
Returns a vector with values that are correlated to vector x.
}
\description{
simulating correlated variable
}
\examples{

cor_var(c(2,4,2,2,6,7,5,6,6,7,9,4,5), 0.5)
}
