% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_data_function.R
\name{gen_param}
\alias{gen_param}
\title{Importing genetic parameters from excel file}
\usage{
gen_param(nameexcelfile)
}
\arguments{
\item{nameexcelfile}{Name or path to the excel file.}
}
\value{
A list with parameters that can be used in the other functions of this package.
}
\description{
This function can be used to import data (heritabilities, variances, correlations) from a excel file with a specific format, and it produces a list. If this list is called 'BPdata', then the genetic parameters do not need to be specified anymore in the other functions.
}
\examples{
\dontrun{
BPdata <- gen_param("example.xlsx")
}
}
